Changes related to getting the Topic/LineItem move-up-and-down functionality working as expected.
Biggest change is the techprofile component now accepts a service that is created in the tech-profile-editor (and elsewhere, still need to test that), as opposed to creating one of its own, and giving access to what its own copy divines. In this way, there will only be one service, and we dont have to worry about them getting out of sync.

Also fixed a bug that wasn't saving Topic sequence info if those topics had no line items associated.